add backend user identification for recipes

+++ backend/src/controllers/recipes.ts
@@ -2,10 +2,17 @@ import { RequestHandler } from "express";
 import RecipeModel from "../models/recipe";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getRecipes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const recipes = await RecipeModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+    const recipes = await RecipeModel.find({
+      userId: authenticatedUserId,
+    }).exec();
+    console.log(recipes)
     res.status(200).json(recipes);
   } catch (error) {
     next(error);
@@ -14,8 +21,11 @@ export const getRecipes: RequestHandler = async (req, res, next) => {
 
 export const getRecipe: RequestHandler = async (req, res, next) => {
   const recipeId = req.params.recipeId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(recipeId)) {
       throw createHttpError(400, "Invalid recipe id");
     }
@@ -24,6 +34,10 @@ export const getRecipe: RequestHandler = async (req, res, next) => {
 
     if (!recipe) {
       throw createHttpError(404, "Recipe not found");
+    }
+
+    if (!recipe.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "Not authorized to access this recipe.");
     }
 
     res.status(200).json(recipe);
@@ -39,7 +53,7 @@ interface CreateRecipeBody {
   portion?: string;
   cooktime?: string;
   imgURL?: string;
-  ingredients?: { amount?: string;  name: string }[];
+  ingredients?: { amount?: string; name: string }[];
   directions?: { text: string }[];
 }
 
@@ -49,6 +63,7 @@ export const createRecipe: RequestHandler<
   CreateRecipeBody,
   unknown
 > = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   const title = req.body.title;
   const description = req.body.description;
   const category = req.body.category;
@@ -59,11 +74,14 @@ export const createRecipe: RequestHandler<
   const directions = req.body.directions;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Recipe must have a title");
     }
 
     const newRecipe = await RecipeModel.create({
+      userId: authenticatedUserId,
       title: title,
       description: description,
       category: category,
@@ -105,6 +123,7 @@ export const updateRecipe: RequestHandler<
   UpdateRecipeBody,
   unknown
 > = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   const recipeId = req.params.recipeId;
   const newTitle = req.body.title;
   const newDescription = req.body.description;
@@ -115,6 +134,8 @@ export const updateRecipe: RequestHandler<
   const newDirections = req.body.directions!;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(recipeId)) {
       throw createHttpError(400, "Invalid recipe id");
     }
@@ -123,6 +144,10 @@ export const updateRecipe: RequestHandler<
 
     if (!recipe) {
       throw createHttpError(404, "Recipe not found");
+    }
+
+    if (!recipe.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "Not authorized to access this recipe.");
     }
 
     recipe.title = newTitle;
@@ -134,7 +159,6 @@ export const updateRecipe: RequestHandler<
     recipe.directions = newDirections;
 
     const updatedRecipe = await recipe.save();
-
     res.status(200).json(updatedRecipe);
   } catch (error) {
     next(error);
@@ -142,9 +166,12 @@ export const updateRecipe: RequestHandler<
 };
 
 export const deleteRecipe: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   const recipeId = req.params.recipeId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(recipeId)) {
       throw createHttpError(400, "Invalid recipe id");
     }
@@ -153,6 +180,10 @@ export const deleteRecipe: RequestHandler = async (req, res, next) => {
 
     if (!recipe) {
       throw createHttpError(404, "Recipe not found");
+    }
+
+    if (!recipe.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "Not authorized to access this recipe.");
     }
 
     await recipe.deleteOne();
